+++ .github/CHANGELOG.md
@@ -310,12 +310,16 @@ random_mat2 = rng.standard_normal(3, requires_grad=False)
 
 <h3>Improvements</h3>
 
+* The adjoint jacobian differentiation method reuses the state computed on the forward pass.
+  This can be turned off to save memory with the Torch and TensorFlow interfaces by passing
+  `adjoint_cache=False` during QNode creation.
+  [(#1341)](https://github.com/PennyLaneAI/pennylane/pull/1341)
+
 * The `Operator` (and by inheritance, the `Operation` and `Observable` class and their children) 
   now have an `id` attribute, which can mark an operator in a circuit, for example to
   identify it on the tape by a tape transform.
   [(#1377)](https://github.com/PennyLaneAI/pennylane/pull/1377)
 
-  
 * Implement special handling for measuring the variance of Projector observables to improve memory usage.
   [(#1368)](https://github.com/PennyLaneAI/pennylane/pull/1368)